requirements.txt
flask
ruamel.yaml
flask-socketio
requests